open gov refactoring - overview component
